Important commit wrt. "BigDecimal bug" (read descr.)
Mark down exhaustively diff actions with same signature within the same plan, otherwise rescheduling won't work!!! Using an instance counter to distinguish and match them appropriately.
Specifically, there was the need to put it not only in the action instante, but in the sibling referenced within the parent reference of the s.i.a. , otherwise it'd have been instantiated with the default 1 value and the scheduling would have generate impossible to produce (correctly). We really hope that this solve the error for our mental sanity. 
Thank you, 
xxx 
phd candidate Devis